Readme beginning

+++ modal/deleteimage.php
@@ -1,13 +1,17 @@
 <?php
 	set_include_path ("../");
-	require 'config/setup.php';
-	$statement = $conn->prepare("DELETE FROM `likes` WHERE postID = :post");
+	require 'config/database.php';
+	$statement = $conn->prepare("SELECT `path` FROM `posts` WHERE id = :post");
 	$statement->bindParam(":post", $_POST['id']);
 	$statement->execute();
-	$statement = $conn->prepare("DELETE FROM `comments` WHERE postID = :post");
-	$statement->bindParam(":post", $_POST['id']);
-	$statement->execute();
-	$statement = $conn->prepare("DELETE FROM `posts` WHERE id = :post");
-	$statement->bindParam(":post", $_POST['id']);
-	$statement->execute();
+
+	$row = $statement->fetch(PDO::FETCH_ASSOC);
+	if ($row)
+	{
+		unlink('../uploads/'.$row['path']);
+
+		$statement = $conn->prepare("DELETE FROM `posts` WHERE id = :post");
+		$statement->bindParam(":post", $_POST['id']);
+		$statement->execute();
+	}
 ?>
